app/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
transformers==4.35.2
torch==2.1.0
numpy==1.24.3
pydantic==1.10.8
peft==0.6.0
python-multipart==0.0.6
pynacl==1.5.0
aiofiles==23.1.0
google-cloud-storage==2.11.0
requests==2.31.0
loguru==0.7.0
accelerate==0.24.1
bitsandbytes==0.41.0
sentencepiece==0.1.99
pillow==10.1.0
celery==5.3.6
redis==5.0.1
hmac==20101010
